Fix Resize reference op on 32-bit platforms
`np.take(data, clamped, axis=axis)` fails on 32-bit Python (e.g. Windows
x86) with `TypeError: Cannot cast array data from dtype('int64') to
dtype('int32') according to the rule 'safe'`, because `clamped` is int64
while NumPy's default index dtype on 32-bit is int32. Cast indices to
`np.intp` so the kernel works on both 32-bit and 64-bit builds.

Regression from #7920.

+++ onnx/reference/ops/op_resize.py
@@ -452,7 +452,9 @@ def _interpolate_1d_along_axis(
         coeffs = coeffs / coeff_sum
 
     # Edge-padding is equivalent to clamping indices into the valid range.
-    clamped = np.clip(neighbor_idxes, 0, input_width - 1)
+    # Cast to intp so np.take works on 32-bit platforms, where the default
+    # integer dtype is int32 and a safe cast from int64 would be rejected.
+    clamped = np.clip(neighbor_idxes, 0, input_width - 1).astype(np.intp)
     gathered = np.take(data, clamped, axis=axis)
 
     coeff_shape = (1,) * axis + coeffs.shape + (1,) * (data.ndim - axis - 1)
